Remove code pertaining to Lasso < 2.4 (released 2014)
Having many ifdefs makes the code less maintainable and also keeps
code that is in practice untested around, because the developers
never compile the code with ancient lasso versions. Confusing
bugs arise because some configuration options only work with some
of the code and not the other.

This removes all code not satisfying the presence of:
- lasso_server_new_from_buffers,
- lasso_server_load_metadata,
- lasso_profile_set_signature_verify_hint